+++ tests/test_violations.py
@@ -145,23 +145,19 @@ def foo():
     # type
     ("type", ":type",),  # missing separator ":"
     ("type", ":type:",),  # missing separator name and type
-    ("type", ":type:int",),  # missing parameter name
+    ("type", ":type: int",),  # missing parameter name
     ("type", ":type a:",),  # missing type
-    ("type", ":type a: ",),  # missing type (only spaces)
     # return
     ("return", ":return",),  # missing separator ":" and description
     ("return", ":return:",),  # missing description
-    ("return", ":return: ",),  # missing description (only spaces)
     ("return", ":return value: description",),  # invalid return name (should be empty)
     # rtype
     ("rtype", ":rtype",),  # missing separator ":" and type
     ("rtype", ":rtype:",),  # missing type
-    ("rtype", ":rtype: ",),  # missing type (only spaces)
     ("rtype", ":rtype value: description",),  # invalid return name (should be empty)
     # raises
     ("raises", ":raises",),  # missing separator ":" and exception
     ("raises", ":raises:",),  # missing exception
-    ("raises", ":raises: ",),  # missing exception (only spaces)
 ])
 def test_DOC002_function(section, value, violations):
     content = f'''
@@ -191,13 +187,8 @@ def foo(a):
     # missing variable name without description
     ("ivar", ":ivar foo:",),
     ("cvar", ":cvar foo:",),
-    # missing variable without name  and description (only spaces)
-    ("ivar", ":ivar:  ",),
-    ("cvar", ":cvar:  ",),
     # missing type
     ("vartype", ":vartype a:",),
-    # missing type (only spaces)
-    ("vartype", ":vartype a: ",),
 ])
 def test_DOC002_class(section, value, violations):
     content = f'''
@@ -341,9 +332,7 @@ def foo(a):
 def test_DOC008_multiline_docstring_no_raise(violations, docs):
     content = f'''
 def foo(a):
-    """
-    {docs}
-    """
+    """{docs}"""
 '''
 
     expected = ()
@@ -387,32 +376,43 @@ def foo():
 
 @pytest.mark.parametrize("section, expected_section", [
     # param type
-    (":type  a: str", ":type  a: str"),  # consecutive spaces between type and name
+    (":type  a: str", ":type  a: str"),  # consecutive spaces left side
     (": type a: str", ": type a: str"),  # leading space before type keyword
     (":type a : str", ":type a : str"),  # trailing space after name
     (":type a:  str", ":type a:  str"),  # leading space before type hint
-    (":type  a:  str", ":type  a:  str"),  # consecutive spaces both sides
+    (":type a: str ", ":type a: str "),  # trailing space after type hint
+    (":type a:str", ":type a:str"),  # missing leading space before type hint
     # param
     (":param  str  a: description", ":param  str  a:"),  # consecutive spaces
     (": param str a: description", ": param str a:"),  # leading space before param keyword
     (":param str a : description", ":param str a :"),  # trailing space after name
     (":param  str a: description", ":param  str a:"),  # leading space before type hint
     (":param str  a: description", ":param str  a:"),  # leading space before name
+    (":param str a:description", ":param str a:description"),  # missing leading space before description
+    (":param str a:  description", ":param str a:  description"),  # consecutive leading spaces before description
+    (":param str a: description ", ":param str a: description "),  # Trailing space after description
     # return
     (": return: description", ": return:"),  # leading space before return keyword
     (":return : description", ":return :"),  # trailing space after return keyword
-    # # rtype
+    (":return:description", ":return:description"),  # missing leading space before description
+    (":return:  description", ":return:  description"),  # consecutive leading spaces before description
+    (":return: description ", ":return: description "),  # Trailing space after description
+    # rtype
     (": rtype: int", ": rtype: int"),  # leading space before rtype keyword
     (":rtype : int", ":rtype : int"),  # trailing space after rtype keyword
     (":rtype:  int", ":rtype:  int"),  # leading space before type hint
-    (": rtype :  int", ": rtype :  int"),  # leading and trailing space both sides
-    # # raises
+    (":rtype: int ", ":rtype: int "),  # trailing space after type hint
+    (":rtype:int", ":rtype:int"),  # missing leading space before type hint
+    # raises
     (": raises ValueError: description", ": raises ValueError:"),  # leading ws before raises keyword
     (":raises  ValueError: description", ":raises  ValueError:"),  # consecutive ws after raises keyword
     (":raises ValueError : description", ":raises ValueError :"),  # trailing ws after error
     (":raises ValueError,  KeyError: description", ":raises ValueError,  KeyError:"),  # Consecutive ws after comma
     (":raises ValueError, KeyError : description", ":raises ValueError, KeyError :"),  # trailing ws after last error
     (":raises ValueError  , KeyError: description", ":raises ValueError  , KeyError:"),  # leading ws before comma
+    (":raises ValueError:description", ":raises ValueError:description"),  # missing leading space before description
+    (":raises ValueError:  description", ":raises ValueError:  description"),  # consecutive leading ws before desc
+    (":raises ValueError: description ", ":raises ValueError: description "),  # trailing space after description
 ])
 def test_DOC010_function(section, violations, expected_section):
     content = f'''
@@ -433,13 +433,18 @@ def foo(a):
 
 
 @pytest.mark.parametrize("section", [
-    ":param str a:  foo  bar  ",
-    ":return:  foo  bar  ",
-    ":raises ValueError:  foo  bar  ",
+    # consecutive internal whitespace in descriptions
+    ":param str a: foo  bar",
+    ":return: foo  bar",
+    ":raises ValueError: foo  bar",
+    # starts with 2+ spaces but starting with newline
+    ":param str a: \n  foo bar",
+    ":return: \n  foo bar",
+    ":raises ValueError: \n  foo bar",
 ])
 def test_DOC010_function_ignoring_descriptions_ws(section, violations):
     content = f'''
-def foo(a):
+def foo(a, b):
     """
     Title.
 
@@ -460,14 +465,20 @@ def foo(a):
     (":vartype a : int", ":vartype a : int"),  # trailing space after name
     (":vartype a:  int", ":vartype a:  int"),  # leading space before type hint
     (":vartype  a:  int", ":vartype  a:  int"),  # consecutive spaces both sides
+    (":vartype a: int ", ":vartype a: int "),  # trailing space after type hint
+    (":vartype a:int", ":vartype a:int"),  # missing leading space before type hint
     # ivar
     (":ivar  a: description", ":ivar  a:"),  # consecutive spaces
     (": ivar a: description", ": ivar a:"),  # leading space before ivar keyword
     (":ivar a : description", ":ivar a :"),  # trailing space after name
+    (":ivar a: description ", ":ivar a: description "),  # trailing space after description
+    (":ivar a:description", ":ivar a:description"),  # missing leading space before description
     # cvar
     (":cvar  a: description", ":cvar  a:"),  # consecutive spaces
     (": cvar a: description", ": cvar a:"),  # leading space before cvar keyword
     (":cvar a : description", ":cvar a :"),  # trailing space after name
+    (":cvar a: description ", ":cvar a: description "),  # trailing space after description
+    (":cvar a:description", ":cvar a:description"),  # missing leading space before description
 ])
 def test_DOC010_class(section, expected_section, violations):
     content = f'''
@@ -488,8 +499,12 @@ class Foo:
 
 
 @pytest.mark.parametrize("section", [
-    ":ivar a:  foo  bar  ",
-    ":cvar a:  foo  bar  ",
+    # consecutive internal whitespace
+    ":ivar a: foo   bar",
+    ":cvar a: foo   bar",
+    # starts with 2+ spaces but starting with newline
+    ":ivar a: \n  foo bar",
+    ":cvar a: \n  foo bar",
 ])
 def test_DOC010_class_ignoring_descriptions_ws(section, violations):
     content = f'''